resources/test/tests/add_cleanup_async.html

<!DOCTYPE HTML>
<html>
<head>
<title>Test#add_cleanup with Promise-returning functions</title>
<script src="../../testharness.js"></script>
<script src="../../testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>

<script>
var resolve1, resolve2;
var completeCount = 0;
test(function(t) {
  add_result_callback(function(result_t) {
    if (result_t.name === t.name) {
      completeCount += 1;
    }
  });
  t.add_cleanup(function() {
    return new Promise(function(_resolve) { resolve1 = _resolve; });
  });
  t.add_cleanup(function() {
    return new Promise(function(_resolve) { resolve2 = _resolve; });
  });
}, 'synchronous test with asynchronous cleanup');

async_test(function(t) {
  assert_equals(
    completeCount, 0, 'test is not immediately considered "complete"'
  );
  Promise.resolve().then(t.step_func(function() {
      assert_equals(
        completeCount, 0, 'test is not asynchronously considered "complete"'
      );

      resolve1();

      Promise.resolve().then(t.step_func(function() {
          assert_equals(
            completeCount,
            0,
            'test is not considered "complete" following resolution of ' +
            'first promise'
          );
          resolve2();
          t.done();
        }));
    }));
}, 'testharness should wait for provided Promises to resolve');
</script>
<script type="text/json" id="expected">
{
  "summarized_status": {
    "status_string": "OK",
    "message": null,
    "stack": null
  },
  "summarized_tests": [
    {
      "status_string": "PASS",
      "name": "synchronous test with asynchronous cleanup",
      "stack": null,
      "message": null,
      "properties": {}
    },
    {
      "status_string": "PASS",
      "name": "testharness should wait for provided Promises to resolve",
      "stack": null,
      "message": null,
      "properties": {}
    }
  ],
  "type": "complete"
}
</script>
</body>
</html>
